refactor(change_detector): extracted ChangeDetectorDefinition
